requirements.txt
torch
torchvision
timm
numpy
pandas
scikit-learn
pyyaml
tqdm
matplotlib
Pillow
